Reusing the process transactions method
