+++ src/app/app/details/[id]/components/RedirectUriList/RedirectUriList.tsx
@@ -9,12 +9,10 @@ import { LoadingModal, LoadingProps } from '@/components/LoadingModal';
 import { Table } from '@/components/Table';
 import { TeamRoles } from '@/types/team';
 import { Toggle } from '@/components/Toggle';
-import { useGlobalAccount, useOnboarding } from '@/hooks';
+import { useContractGA, useGlobalAccount, useOnboarding } from '@/hooks';
 import { useSession } from 'next-auth/react';
 
 import configuration from '@/config';
-import { IKernelOperationStatus } from '@/types/wallet';
-import { bundlerActions, ENTRYPOINT_ADDRESS_V07 } from 'permissionless';
 import DimoLicenseABI from '@/contracts/DimoLicenseContract.json';
 
 interface IProps {
@@ -26,7 +24,8 @@ export const RedirectUriList: FC<IProps> = ({ list = [], refreshData }) => {
   const [isOpened, setIsOpened] = useState<boolean>(false);
   const [loadingStatus, setLoadingStatus] = useState<LoadingProps>();
   const { workspace } = useOnboarding();
-  const { organizationInfo, getKernelClient } = useGlobalAccount();
+  const { organizationInfo } = useGlobalAccount();
+  const { processTransactions } = useContractGA();
   const { data: session } = useSession();
   const { user: { role = '' } = {} } = session ?? {};
 
@@ -49,28 +48,6 @@ export const RedirectUriList: FC<IProps> = ({ list = [], refreshData }) => {
       }
     }];
     await processTransactions(transaction);
-  };
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const processTransactions = async (transactions: Array<any>) => {
-    if (!organizationInfo) return {} as IKernelOperationStatus;
-    const kernelClient = await getKernelClient(organizationInfo);
-    const dcxExchangeOpHash = await kernelClient.sendUserOperation({
-      userOperation: {
-        callData: await kernelClient.account.encodeCallData(transactions),
-      },
-    });
-
-    const bundlerClient = kernelClient.extend(
-      bundlerActions(ENTRYPOINT_ADDRESS_V07),
-    );
-
-    const { reason } =
-      await bundlerClient.waitForUserOperationReceipt({
-        hash: dcxExchangeOpHash,
-      });
-
-    if (reason) return Promise.reject(reason);
   };
 
   const renderToggleStatus = ({ id, uri, status }: IRedirectUri) => {
